PHPUnit 9.5 tests work, added AssertArraySubet trait for backwards compatability with our PHPUnit tests

+++ src/Response/AddApiKeyResponse.php
@@ -50,7 +50,7 @@ final class AddApiKeyResponse extends AbstractResponse
 
             $retry++;
             $factor = ceil($retry / 10);
-            usleep($factor * $time); // 0.1 second
+            usleep((int) ($factor * $time)); // 0.1 second
         } while (true);
     }
 }
