+++ src/Response/DeleteApiKeyResponse.php
@@ -53,7 +53,7 @@ final class DeleteApiKeyResponse extends AbstractResponse
 
             $retry++;
             $factor = ceil($retry / 10);
-            usleep($factor * $time); // 0.1 second
+            usleep((int) ($factor * $time)); // 0.1 second
         } while (true);
     }
 }
